Refactor: Split LNB components
