+++ src/styles/global.scss
@@ -24,3 +24,17 @@
   line-height: 19px;
   color: colors.$GRAY800;
 }
+
+.main-text {
+  font-size: 16px;
+  font-weight: 700;
+  line-height: 19px;
+  color: colors.$GRAY800;
+}
+
+.sub-text {
+  font-size: 12px;
+  font-weight: 500;
+  line-height: 14px;
+  color: colors.$GRAY300;
+}
